code/train/requirements.txt
tensorflow-transform==0.13.0
apache-beam[gcp]==2.11.0
sklearn